Add a quick version check to avoid extra restore.
RestoreCoreAsync might receive old data flow versions which should not trigger nugert restores.

This change checks compares only the saved version of the active configured project with the one received in RestoreCoreAsync and if it is older than the current saved version we can skip this Nuget restore because it was already proccessed.